+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 ### Improvements
 - Use multicolor channel icons for better appearance in the dark mode
 - Add SITE_LOGO_URL setting (#323)
+- Add admin action to log in as any user
 
 ### Bug Fixes
 - Fix dark mode styling issues in Cron Syntax Cheatsheet